+++ src/writer/json.rs
@@ -12,11 +12,19 @@
 //!
 //! [1]: https://github.com/cucumber/cucumber-json-schema
 
-use std::{fmt::Debug, io, time::SystemTime};
+use std::{
+    fmt::{self, Debug},
+    io, mem,
+    time::SystemTime,
+};
 
 use async_trait::async_trait;
+use base64::Engine as _;
+use derive_more::Display;
 use inflector::Inflector as _;
-use serde::Serialize;
+use mime::Mime;
+use once_cell::sync::Lazy;
+use serde::{Serialize, Serializer};
 
 use crate::{
     cli, event,
@@ -58,6 +66,12 @@ pub struct Json<Out: io::Write> {
     /// [`Scenario`]: gherkin::Scenario
     /// [`Hook`]: event::Hook
     started: Option<SystemTime>,
+
+    /// [`event::Scenario::Log`]s of current [`Hook`] or [`Step`].
+    ///
+    /// [`Scenario`]: gherkin::Scenario
+    /// [`Hook`]: event::Hook
+    logs: Vec<String>,
 }
 
 #[async_trait(?Send)]
@@ -155,6 +169,7 @@ impl<Out: io::Write> Json<Out> {
             output,
             features: Vec::new(),
             started: None,
+            logs: Vec::new(),
         }
     }
 
@@ -170,6 +185,7 @@ impl<Out: io::Write> Json<Out> {
         use event::Scenario;
 
         match ev {
+            Scenario::Started => {}
             Scenario::Hook(ty, ev) => {
                 self.handle_hook_event(feature, rule, scenario, ty, ev, meta);
             }
@@ -189,8 +205,12 @@ impl<Out: io::Write> Json<Out> {
                     feature, rule, scenario, "scenario", &st, ev, meta,
                 );
             }
-            // TODO: report logs for each `Scenario`.
-            Scenario::Started | Scenario::Finished | Scenario::Log(_) => {}
+            Scenario::Log(msg) => {
+                self.logs.push(msg);
+            }
+            Scenario::Finished => {
+                self.logs.clear();
+            }
         }
     }
 
@@ -233,6 +253,10 @@ impl<Out: io::Write> Json<Out> {
                     duration: duration(),
                     error_message: None,
                 },
+                embeddings: mem::take(&mut self.logs)
+                    .into_iter()
+                    .map(Embedding::from_log)
+                    .collect(),
             },
             Hook::Failed(_, info) => HookResult {
                 result: RunResult {
@@ -240,6 +264,10 @@ impl<Out: io::Write> Json<Out> {
                     duration: duration(),
                     error_message: Some(coerce_error(&info).into_owned()),
                 },
+                embeddings: mem::take(&mut self.logs)
+                    .into_iter()
+                    .map(Embedding::from_log)
+                    .collect(),
             },
         };
 
@@ -316,14 +344,19 @@ impl<Out: io::Write> Json<Out> {
             },
         };
 
-        let el = self.mut_or_insert_element(feature, rule, scenario, ty);
-        el.steps.push(Step {
+        let step = Step {
             keyword: step.keyword.clone(),
             line: step.position.line,
             name: step.value.clone(),
             hidden: false,
             result,
-        });
+            embeddings: mem::take(&mut self.logs)
+                .into_iter()
+                .map(Embedding::from_log)
+                .collect(),
+        };
+        let el = self.mut_or_insert_element(feature, rule, scenario, ty);
+        el.steps.push(step);
     }
 
     /// Inserts the given `scenario`, if not present, and then returns a mutable
@@ -367,6 +400,66 @@ impl<Out: io::Write> Json<Out> {
                 f.elements.len() - 1
             });
         f.elements.get_mut(el_pos).unwrap_or_else(|| unreachable!())
+    }
+}
+
+/// [`base64`] encoded data.
+#[derive(Clone, Debug, Display, Serialize)]
+#[serde(transparent)]
+pub struct Base64(String);
+
+impl Base64 {
+    /// Used [`base64::engine`].
+    const ENGINE: base64::engine::GeneralPurpose =
+        base64::engine::general_purpose::STANDARD;
+
+    /// Encodes `bytes` as [`base64`].
+    pub fn encode(bytes: impl AsRef<[u8]>) -> Self {
+        Self(Self::ENGINE.encode(bytes))
+    }
+
+    /// Decodes this [`base64`] encoded data.
+    #[must_use]
+    pub fn decode(&self) -> Vec<u8> {
+        Self::ENGINE.decode(&self.0).unwrap_or_else(|_| {
+            unreachable!(
+                "Only way to construct this type is `Base64::encode`, so \
+                 should contain valid `base64` encoded `String`",
+            )
+        })
+    }
+}
+
+/// Embedded data.
+#[derive(Clone, Debug, Serialize)]
+pub struct Embedding {
+    /// [`base64`] encoded data.
+    pub data: Base64,
+
+    /// [`Mime`] of this [`Embedding::data`].
+    #[serde(serialize_with = "serialize_display")]
+    pub mime_type: Mime,
+
+    /// Optional name of the [`Embedding`].
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub name: Option<String>,
+}
+
+impl Embedding {
+    /// Creates [`Embedding`] from [`event::Scenario::Log`].
+    fn from_log(msg: impl AsRef<str>) -> Self {
+        /// [`Mime`] of the [`event::Scenario::Log`] [`Embedding`].
+        static LOG_MIME: Lazy<Mime> = Lazy::new(|| {
+            "text/x.cucumber.log+plain"
+                .parse()
+                .unwrap_or_else(|_| unreachable!("valid MIME"))
+        });
+
+        Self {
+            data: Base64::encode(msg.as_ref()),
+            mime_type: LOG_MIME.clone(),
+            name: None,
+        }
     }
 }
 
@@ -445,6 +538,18 @@ pub struct Step {
 
     /// [`RunResult`] of this [`Step`].
     pub result: RunResult,
+
+    /// [`Embedding`]s of this [`Step`].
+    ///
+    /// Although this field isn't present in [JSON schema][1], all major
+    /// implementations have it: [Java], [JavaScript], [Ruby]
+    ///
+    /// [1]: https://github.com/cucumber/cucumber-json-schema
+    /// [Java]: https://bit.ly/3J66vxT
+    /// [JavaScript]: https://bit.ly/41HSTAf
+    /// [Ruby]: https://bit.ly/3kAJRof
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub embeddings: Vec<Embedding>,
 }
 
 /// [`Serialize`]able result of running a [`Before`] or [`After`] hook.
@@ -455,6 +560,19 @@ pub struct Step {
 pub struct HookResult {
     /// [`RunResult`] of the hook.
     pub result: RunResult,
+
+    /// [`Embedding`]s of this [`Hook`].
+    ///
+    /// Although this field isn't present in [JSON schema][1], all major
+    /// implementations have it: [Java], [JavaScript], [Ruby]
+    ///
+    /// [`Hook`]: event::Hook
+    /// [1]: https://github.com/cucumber/cucumber-json-schema
+    /// [Java]: https://bit.ly/3J66vxT
+    /// [JavaScript]: https://bit.ly/41HSTAf
+    /// [Ruby]: https://bit.ly/3kAJRof
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub embeddings: Vec<Embedding>,
 }
 
 /// [`Serialize`]able [`gherkin::Background`] or [`gherkin::Scenario`].
@@ -630,6 +748,7 @@ impl Feature {
                         duration: 0,
                         error_message: Some(err.to_string()),
                     },
+                    embeddings: Vec::new(),
                 }],
             }],
         }
@@ -672,6 +791,7 @@ impl Feature {
                         duration: 0,
                         error_message: Some(err.to_string()),
                     },
+                    embeddings: Vec::new(),
                 }],
             }],
         }
@@ -692,4 +812,14 @@ impl PartialEq<gherkin::Feature> for Feature {
             .unwrap_or_default()
             && self.name == feature.name
     }
+}
+
+/// Helper to  use `#[serde(serialize_with = serialize_display)]` with any type
+/// that implements [`fmt::Display`].
+fn serialize_display<T, S>(display: &T, ser: S) -> Result<S::Ok, S::Error>
+where
+    T: fmt::Display,
+    S: Serializer,
+{
+    format_args!("{display}").serialize(ser)
 }