+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 ### BC Breaks
 
 - Added `Log` variant to `event::Scenario`. ([#258])
+- Added `embeddings` field to `writer::json::Step` and `writer::json::HookResult`. ([#261])
 
 ### Added
 
@@ -21,6 +22,7 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 
 [#213]: /../../issues/213
 [#258]: /../../pull/258
+[#261]: /../../pull/261